Move xdr2json into its own library and add shared FFI
