+++ cmd/soroban-rpc/lib/preflight/Cargo.toml
@@ -1,20 +1,23 @@
 [package]
 name = "preflight"
-version = "21.3.0"
+version = "21.4.0"
 publish = false
 
 [lib]
 crate-type = ["staticlib"]
 
 [dependencies]
+ffi = { path = "../ffi" }
+
 base64 = { workspace = true }
 libc = { workspace = true }
 sha2 = { workspace = true }
+
 # we need the testutils feature in order to get backtraces in the preflight library
 # when soroban rpc is configured to run with --preflight-enable-debug
 soroban-env-host = { workspace = true, features = ["recording_mode", "testutils", "unstable-next-api"]}
 soroban-simulation = { workspace = true }
-stellar-xdr = { workspace = true }
+
 anyhow = { workspace = true }
-serde_json = { workspace = true }
 rand = { workspace = true }
+serde_json = { workspace = true }
