+++ cmd/soroban-rpc/lib/shared.h
@@ -1,0 +1,4 @@
+typedef struct xdr_t {
+    unsigned char *xdr;
+    size_t        len;
+} xdr_t;
