+++ cmd/soroban-rpc/lib/xdr2json.h
@@ -1,3 +1,5 @@
+#include "shared.h"
+
 typedef struct {
     const char* const json;
     const char* const error;
