+++ cmd/soroban-rpc/lib/xdr2json/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "xdr2json"
+version = "21.0.0"
+publish = false
+
+[lib]
+crate-type = ["staticlib"]
+
+[dependencies]
+ffi = { path = "../ffi" }
+libc = { workspace = true }
+sha2 = { workspace = true }
+base64 = { workspace = true }
+anyhow = { workspace = true }
+serde_json = { workspace = true }
+rand = { workspace = true }
+
+stellar-xdr = { workspace = true }
+soroban-env-host = { workspace = true, features = ["unstable-next-api"]}
