+++ cmd/soroban-rpc/lib/xdr2json/src/lib.rs
@@ -1,0 +1,107 @@
+extern crate ffi;
+extern crate soroban_env_host;
+extern crate stellar_xdr;
+
+use std::{panic, str::FromStr};
+
+use anyhow::Result;
+use ffi::*;
+use soroban_env_host::{xdr, DEFAULT_XDR_RW_LIMITS};
+
+#[repr(C)]
+pub struct ConversionResult {
+    json: *mut libc::c_char,
+    error: *mut libc::c_char,
+}
+
+struct RustConversionResult {
+    json: String,
+    error: String,
+}
+
+/// Takes in a string name of an XDR type in the Stellar Protocol (i.e. from the
+/// `stellar_xdr` crate) as well as a raw byte structure and returns a structure
+/// containing the JSON-ified string of the given structure.
+///
+/// # Errors
+///
+/// On error, the struct's `error` field will be filled out with the appropriate
+/// message that caused the function to panic.
+///
+/// # Panics
+///
+/// This should never panic due to `catch_json_to_xdr_panic` catching and
+/// unwinding all panics to stringified error messages.
+#[no_mangle]
+pub extern "C" fn xdr_to_json(typename: *mut libc::c_char, xdr: CXDR) -> *mut ConversionResult {
+    let result = catch_json_to_xdr_panic(Box::new(move || {
+        let type_str = from_c_string(typename);
+        let the_type = xdr::TypeVariant::from_str(&type_str).unwrap();
+
+        let xdr_bytearray = from_c_xdr(xdr);
+        let mut buffer = xdr::Limited::new(xdr_bytearray.as_slice(), DEFAULT_XDR_RW_LIMITS.clone());
+
+        let t = xdr::Type::read_xdr_to_end(the_type, &mut buffer).unwrap();
+
+        Ok(RustConversionResult {
+            json: serde_json::to_string(&t).unwrap(),
+            error: String::new(),
+        })
+    }));
+
+    // Caller is responsible for calling free_conversion_result.
+    Box::into_raw(Box::new(ConversionResult {
+        json: ffi::string_to_c(result.json),
+        error: string_to_c(result.error),
+    }))
+}
+
+/// Frees memory allocated for the corresponding conversion result.
+///
+/// # Safety
+///
+/// You should *only* use this to free the return value of `xdr_to_json`.
+#[no_mangle]
+pub unsafe extern "C" fn free_conversion_result(ptr: *mut ConversionResult) {
+    if ptr.is_null() {
+        return;
+    }
+
+    unsafe {
+        free_c_string((*ptr).json);
+        free_c_string((*ptr).error);
+        drop(Box::from_raw(ptr));
+    }
+}
+
+/// Runs a JSON conversion operation and unwinds panics.
+///
+/// It is modeled after `catch_preflight_panic()` and will always return valid
+/// JSON in the result's `json` field and an error string in `error` if a panic
+/// occurs.
+fn catch_json_to_xdr_panic(
+    op: Box<dyn Fn() -> Result<RustConversionResult>>,
+) -> RustConversionResult {
+    // catch panics before they reach foreign callers (which otherwise would result in
+    // undefined behavior)
+    let res: std::thread::Result<Result<RustConversionResult>> =
+        panic::catch_unwind(panic::AssertUnwindSafe(op));
+
+    match res {
+        Err(panic) => match panic.downcast::<String>() {
+            Ok(panic_msg) => RustConversionResult {
+                json: "{}".to_string(),
+                error: format!("xdr_to_json() failed: {panic_msg}"),
+            },
+            Err(_) => RustConversionResult {
+                json: "{}".to_string(),
+                error: "xdr_to_json() failed: unknown cause".to_string(),
+            },
+        },
+        // See https://docs.rs/anyhow/latest/anyhow/struct.Error.html#display-representations
+        Ok(r) => r.unwrap_or_else(|e| RustConversionResult {
+            json: "{}".to_string(),
+            error: format!("{e:?}"),
+        }),
+    }
+}
